+++ app/commands/characters_context/dc20/create_command.rb
@@ -32,7 +32,7 @@ module CharactersContext
       def do_persist(input)
         character = ::Dc20::Character.create!(input.slice(:user, :name, :data))
 
-        attach_feats(character, input[:ancestry_feats].values.flatten)
+        attach_feats(character, input[:main_class], input[:ancestry_feats].values.flatten)
 
         if input[:avatar_file]
           ImageProcessingContext::AttachAvatarByFileJob.perform_later(character_id: character.id, file: input[:avatar_file])
@@ -49,8 +49,8 @@ module CharactersContext
           .then { |result| Dc20Character::ClassBuilder.new.call(result: result) }
       end
 
-      def attach_feats(character, feat_slugs)
-        feats_for_adding = ::Dc20::Feat.where(origin: 0, slug: feat_slugs).map do |feat|
+      def attach_feats(character, main_class, feat_slugs)
+        feats_for_adding = feats_relation(main_class, feat_slugs).map do |feat|
           {
             feat_id: feat.id,
             character_id: character.id,
@@ -60,6 +60,11 @@ module CharactersContext
           }
         end
         ::Character::Feat.upsert_all(feats_for_adding) if feats_for_adding.any?
+      end
+
+      def feats_relation(main_class, feat_slugs)
+        ::Dc20::Feat.where(origin: 0, slug: feat_slugs)
+          .or(::Dc20::Feat.where(origin: [1, 2], origin_value: main_class).where("conditions ->> 'level' = '1'"))
       end
     end
   end