+++ db/migrate/20251107071838_add_price_to_feats.rb
@@ -1,0 +1,7 @@
+class AddPriceToFeats < ActiveRecord::Migration[8.1]
+  def change
+    add_column :feats, :price, :jsonb, default: {}, comment: 'Цена активации способности'
+
+    Feat.update_all(price: {})
+  end
+end